revising doc generator and adding test types
