+++ docs/standard-type-reference.md
@@ -1,0 +1,41 @@
+---
+id: standard-type-reference
+title: "Standard Type Reference"
+---
+# Standard Type Reference
+
+ZIO Schema provides a number of built-in primitive types, that we can use to represent our data. These can be seen in the following table:
+
+|Standard Type|JVM Support|ScalaJS Support|Scala Native Support|
+|--------------|:--------------:|:--------------:|:--------------:|
+|`Boolean`|✅|✅|✅|
+|`Byte`|✅|✅|✅|
+|`Char`|✅|✅|✅|
+|`Chunk[Byte]`|✅|✅|✅|
+|`Double`|✅|✅|✅|
+|`Float`|✅|✅|✅|
+|`Int`|✅|✅|✅|
+|`Long`|✅|✅|✅|
+|`Short`|✅|✅|✅|
+|`String`|✅|✅|✅|
+|`Unit`|✅|✅|✅|
+|`java.math.BigDecimal`|✅|✅|✅|
+|`java.math.BigInteger`|✅|✅|✅|
+|`java.time.DayOfWeek`|✅|✅|✅|
+|`java.time.Duration`|✅|✅|✅|
+|`java.time.Instant`|✅|✅|✅|
+|`java.time.LocalDate`|✅|✅|✅|
+|`java.time.LocalDateTime`|✅|✅|✅|
+|`java.time.LocalTime`|✅|✅|✅|
+|`java.time.Month`|✅|✅|✅|
+|`java.time.MonthDay`|✅|✅|✅|
+|`java.time.OffsetDateTime`|✅|✅|✅|
+|`java.time.OffsetTime`|✅|✅|✅|
+|`java.time.Period`|✅|✅|✅|
+|`java.time.Year`|✅|✅|✅|
+|`java.time.YearMonth`|✅|✅|✅|
+|`java.time.ZoneId`|✅|✅|✅|
+|`java.time.ZoneOffset`|✅|✅|✅|
+|`java.time.ZonedDateTime`|✅|✅|✅|
+|`java.util.Currency`|✅|❌|❌|
+|`java.util.UUID`|✅|❌|✅|
