+++ zio-schema-json/jvm/src/test/scala-2/zio/schema/codec/JsonPlatformSpecific.scala
@@ -2,12 +2,11 @@ package zio.schema.codec
 
 import zio.schema.codec.JsonCodecSpec.{ assertDecodes, assertEncodesJson, stringify, test }
 import zio.schema.{ Schema, StandardType }
-import zio.test.{ Gen, check }
+import zio.test.{ Gen, Spec, TestEnvironment, check }
 
 object JsonPlatformSpecific {
 
-  @SuppressWarnings(Array("scalafix:ExplicitResultTypes"))
-  val platformSpecificEncoderTests = Seq(
+  val platformSpecificEncoderTests: List[Spec[TestEnvironment, Nothing]] = List(
     test("Currency") {
       assertEncodesJson(
         Schema.Primitive(StandardType.CurrencyType),
@@ -16,8 +15,7 @@ object JsonPlatformSpecific {
     }
   )
 
-  @SuppressWarnings(Array("scalafix:ExplicitResultTypes"))
-  val platformSpecificDecoderTests = Seq(
+  val platformSpecificDecoderTests: List[Spec[TestEnvironment, DecodeError]] = List(
     test("Currency") {
       check(Gen.currency)(
         currency => assertDecodes(Schema[java.util.Currency], currency, stringify(currency.getCurrencyCode))
